requirements.txt
asgiref==3.9.2
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
djangorestframework==3.16.1
greenlet==3.2.4
llvmlite==0.44.0
narwhals==2.5.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.71b0
plotly==6.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
setuptools==78.1.1
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
gunicorn==23.0.0
psycopg2-binary==2.9.9

